With/SDR_Bot.py/clint-sdr-bot/requirements.txt
openai>=1.3.8
PyPDF2
python-dotenv
pydub
requests
whisper